refactor: move common const and types to `@/lib/const`
